0.0.76-1 Pre-alpha - Fixed notification when trying to change the rating of a comment again; - Now the comment filter ignores the review of comments by moderation and site administration; - Corrected numbering of replies to comments; - It is now impossible to delete primary user groups and the primary user that are generated during system installation; - Fixed display of user group name in comments to posts;

+++ api/usersGroup/delete.handler.php
@@ -28,12 +28,17 @@ if ($system_core->client->is_logged(2)) {
       if (UserGroup::exists_by_id($system_core, $user_group_id)) {
         $user_group = new UserGroup($system_core, $user_group_id);
 
-        $user_group_is_deleted = $user_group->delete();
-        if ($user_group_is_deleted) {
-          $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
-          $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+        if ($user_group_id > 4) {
+          $user_group_is_deleted = $user_group->delete();
+          if ($user_group_is_deleted) {
+            $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
+            $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+          } else {
+            $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+            $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+          }
         } else {
-          $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+          $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USERS_GROUP_ERROR_DELETION_PROHIBITED')) : $handler_message;
           $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
         }
       } else {
